Elections list

+++ client/src/components/Election/components/VotingBooth/index.jsx
@@ -10,8 +10,7 @@ import {
 } from 'firebase/database'
 import {
   DbContext,
-  SecretContext,
-  UserAddCandidateContext,
+  UserContext,
 } from 'contexts'
 import {
   Panel,
@@ -55,8 +54,7 @@ function VotingBooth({
   const [candidateDeleteWarningModalOpen, setCandidateDeleteWarningModalOpen] = useState(false)
   const [dontAskDeletionConfirmation, setDontAskDeletionConfirmation] = useState(localStorage.getItem('dont_ask_deletion_confirmation') === 'true')
   const db = useContext(DbContext)
-  const secret = useContext(SecretContext)
-  const addCandidateId = useContext(UserAddCandidateContext)
+  const userContext = useContext(UserContext)
   const users = []
   Object.keys(election.users).forEach((key) => {
     const u = election.users[key]
@@ -87,6 +85,8 @@ function VotingBooth({
   }, [db, election, user])
 
   const handleChangeVote = (newVote, updateHasVoted = true) => {
+    if (!user.id && userContext.id === process.env.REACT_APP_SUPER_USER_ID) return
+
     setVote(newVote)
     set(ref(db, `votes/${election.fullId}/${user.fullId}`), newVote)
     if (updateHasVoted) {
@@ -97,12 +97,16 @@ function VotingBooth({
   }
 
   const handleCastVote = () => {
+    if (!user.id && userContext.id === process.env.REACT_APP_SUPER_USER_ID) return
+
     update(ref(db, `elections/${election.fullId}/users/${user.fullId}`), {
       hasVoted: true,
     })
   }
 
   const handleCountVotes = () => {
+    if (userContext.id === process.env.REACT_APP_SUPER_USER_ID) return
+
     setClickedCountVotes(true)
     update(ref(db, `elections/${election.fullId}`), {
       isFinished: true,
@@ -113,7 +117,7 @@ function VotingBooth({
           electionId: election.fullId,
           user: {
             id: user.id,
-            secret,
+            secret: userContext.secret,
           },
         })
           .then(({ status }) => {
@@ -130,11 +134,12 @@ function VotingBooth({
   }
 
   const userCanDeleteCandidate = (candidateId) => {
+    if (!user.id && userContext.id === process.env.REACT_APP_SUPER_USER_ID) return false
     if (get(election, 'candidates.length', 0) <= 1) return false
 
     if (
       user.id === election.creator
-      || addCandidateId === get(find(
+      || userContext.addCandidateId === get(find(
         get(election, 'candidates', []),
         (c) => c.id === candidateId,
       ), 'addedBy')
@@ -146,6 +151,7 @@ function VotingBooth({
   }
 
   const deleteCandidate = (candidateId) => {
+    if (!user.id && userContext.id === process.env.REACT_APP_SUPER_USER_ID) return
     localStorage.setItem('dont_ask_deletion_confirmation', dontAskDeletionConfirmation)
 
     const deleteId = candidateId || candidateToBeDeleted
@@ -170,6 +176,7 @@ function VotingBooth({
   }
 
   const handleDeleteCandidate = (candidateId) => {
+    if (!user.id && userContext.id === process.env.REACT_APP_SUPER_USER_ID) return
     if (localStorage.getItem('dont_ask_deletion_confirmation') === 'true') {
       deleteCandidate(candidateId)
       return
@@ -183,6 +190,7 @@ function VotingBooth({
   }
 
   const handleAddCandidate = (candidateName) => {
+    if (!user.id && userContext.id === process.env.REACT_APP_SUPER_USER_ID) return
     const similarityAllowed = get(election, 'candidateNameSimilarityAllowed', 0)
     const candidateNames = map(
       get(election, 'candidates', []),
@@ -217,7 +225,7 @@ function VotingBooth({
 
     const newCandidates = cloneDeep(get(election, 'candidates', []))
     newCandidates.push({
-      addedBy: addCandidateId,
+      addedBy: userContext.addCandidateId,
       name: candidateName,
       id: generateCandidateId(election.candidates),
     })
